requirements.txt
# Core Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0

# Testing (Optional)
pytest>=7.4.0
